pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "asyncio>=4.0.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qwen-vl-utils>=0.0.14",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
    "vllm>=0.15.1",
]
